[STF] Fix and rename ContextMenuShareWithTest
The current context menu entry simply opens
the sharing dialog. Therefore the test is renamed
and only checks that the dialog is opened.
